Move the header files into a libEMF subdirectory.
Update one of the test EMFs to account for the extra words in the
extended EMF header.

I updated the autoconf stuff to Fedora 11 to make it build here. Not
sure if that's an improvement. Also, I had to modify
/usr/share/aclocal/libtool.m4 to get around an error (../libtool:
fork: retry: Resource temporarily unavailable).

Next step: Convert to CMake!



git-svn-id: https://svn.code.sf.net/p/libemf/code/trunk@47 07b56ee9-b158-4c8b-9baa-ff89d65b121f